Pass user choices as a context
