+++ src/frontend/packages/core/src/components/livekit/LiveKitMeeting/index.tsx
@@ -1,40 +1,26 @@
-import { LiveKitRoom, VideoConference, PreJoin} from '@livekit/components-react'
+import { LiveKitRoom, VideoConference, PreJoin, useTracks, useMediaDevices, useRoomContext } from '@livekit/components-react'
 
 import { DEFAULT_LIVEKIT_DOMAIN } from '../../../utils/settings'
 
 import '@livekit/components-styles';
 import { Room } from 'livekit-client';
 import { Fragment, useMemo, useState } from 'react';
-import { Box } from 'grommet';
+import { Box, Button } from 'grommet';
 import { useAuthContext } from '../../../context';
 import { useNavigate } from 'react-router-dom';
+import { usePresets } from '../../../views/rooms/livekit';
 
 export interface LiveKitMeetingProps {
     token: string
-}
-
-interface LocalUserChoices {
-    videoEnabled: boolean,
-    audioEnabled: boolean,
-    videoDeviceId: string,
-    audioDeviceId: string,
-    username: string,
 }
 
 export const LiveKitMeeting = ({
     ...props
 }: LiveKitMeetingProps) => {
 
-    const { user } = useAuthContext()
     const navigate = useNavigate()
-    const [ready, setReady] = useState<boolean>(false)
-    const [choices, setChoices] = useState<LocalUserChoices>({
-        videoEnabled: true,
-        audioEnabled: false,
-        videoDeviceId: '',
-        audioDeviceId: '',
-        username: user?.username ?? '',
-    })
+
+    const choices = usePresets()
 
     const roomOptions = useMemo(() => {
         return ({
@@ -44,33 +30,20 @@ export const LiveKitMeeting = ({
             audioCaptureDefaults: {
                 deviceId: choices.audioDeviceId ?? undefined
             },
-            dynacast: true
+            dynacast: true,
         })
     }, [choices])
 
-    const handlePreJoinSubmit = (choices: LocalUserChoices) => {
-        setChoices(choices)
-        setReady(true)
-    }
-
-    const onDisconnected = () => {
+    const handleDisconnect = () => {
         navigate('/')
     }
 
     return (
-        <Fragment>
-            {!ready &&
-                <Box style={{ backgroundColor: "black", width: "100%", height: "100%", display: "flex", justifyContent: "center", alignItems: "center" }}>
-                    <PreJoin style={{ backgroundColor: "black" }} data-lk-theme="default" onSubmit={handlePreJoinSubmit} defaults={choices} persistUserChoices={false}></PreJoin>
-                </Box>
-            }
-            {ready && <LiveKitRoom onDisconnected={onDisconnected} data-lk-theme="default" serverUrl={DEFAULT_LIVEKIT_DOMAIN} token={props.token} connect={true} room={new Room(roomOptions)} audio={choices.audioEnabled} video={choices.videoEnabled} connectOptions={{ autoSubscribe: true }}>
-                <VideoConference />
-            </LiveKitRoom>}
-        </Fragment>
+        <LiveKitRoom data-lk-theme="default" serverUrl={DEFAULT_LIVEKIT_DOMAIN} token={props.token} connect={true} room={new Room(roomOptions)} audio={choices.audioEnabled} onDisconnected={handleDisconnect} video={choices.videoEnabled} connectOptions={{ autoSubscribe: true }}>
+            <VideoConference />
+        </LiveKitRoom>
     )
 }
-
